Switch to cookie for JWT
Initially I had the JWT in the response of signin/login.
However, this requires the client to "store" the JWT somewhere.
I had it stored in a "global" JS variable, but when the page
gets reloaded, it obviously disappears and the user would have to logon
again. Then I tought I could use localStorage but I knew that is
unsecure (also see [this](https://blog.logrocket.com/jwt-authentication-best-practices/#:~:text=Don't%20store%20it%20in,storage%20(or%20session%20storage).&text=The%20JWT%20needs%20to%20be,JavaScript%20running%20in%20the%20browser.)

So then the best option is to just send the JWT in a Cookie, which gets
automatically stored by the browser and (with some withCredentials
wrangling), then will get send with each additional request to the
server.

For now, I am also sending the token in the response body, but it
should not be used anymore (depricated).