Update remoteRoot
In order for the debugger to work properly, it needs to know where the
source code is inside the container.
I used the default generated `launch.json` file which points
`remoteRoot` to `/usr/src/app`. However, if you check the `Dockerfile`
for my app, you can see I am copying my source code to `/opt/app` inside the
container. So I had to set `remoteRoot` to that location instead of the
default one.

Without this, I was unable to place breakpoints in the VS Code Debugger
(although I was able to use the `debugger` statement). The breakpoints
would appear as 'Unbound' and the code just wouldn't stop at the
breakpoints.