Add NOT NULL constraints on FK
I thought that FKs cannot be NULL by default but that is incorrect.
Not only can they be NULL, but if they are NULL and
if they are references in a combined UNIQUE constraints, that
UNIQUE constraint will let any combination with NULL pass, so you can
have duplicates with NULL, e.g. (1, NULL), (1,NULL) is totally fine.
Adding the NULL constraint prevents this because no you cannot have
NULL at all